Compute Cell-Type-Specific Average Expression of Spatial Clusters

Compute Cell-Type-Specific Average Expression of Spatial Clusters

Usage

ComputeAvg(
  normdata,
  scmeta,
  cluster = "SE",
  Region = NULL,
  scale = TRUE,
  ncores = 4
)

Arguments

normdata

Numeric matrix of normalized expression data, where rows represent genes and columns represent cells.

scmeta

Data frame containing metadata associated with each cell, including spatial cluster and cell type annotations.

cluster

Character string specifying the column name in 'scmeta' containing spatial cluster annotations.

Region

Character string specifying the column name in metadata data frames containing region annotations (default: NULL).

scale

A boolean specifying whether to do univariance normalization.

ncores

Integer specifying the number of CPU cores to use for parallel processing.

Value

A matrix of average expression, where rows represent genes and columns represent spatial clusters from the sample.

Examples

library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("13Rc5Rsu8jbnEYYfUse-xQ7ges51LcI7n"),
                    "HumanMelanomaPatient1_subset_counts.tsv.gz",
                    overwrite = TRUE, verbose = FALSE)
drive_download(as_id("1GUbbEBttiu2ntVpaIcIK7YobaKKwbqnM"),
                    "HumanMelanomaPatient1_subset_SpatialEcoTyper_results.rds",
                    overwrite = TRUE, verbose = FALSE)
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
tmpobj <- CreateSeuratObject(scdata) %>%
        SCTransform(clip.range = c(-10, 10), verbose = FALSE)
seurat_version = as.integer(gsub("\\..*", "", as.character(packageVersion("Seurat"))))
if(seurat_version<5){
  normdata <- GetAssayData(tmpobj, "data")
}else{
  normdata <- tmpobj[["SCT"]]$data
}
metadata = readRDS("HumanMelanomaPatient1_subset_SpatialEcoTyper_results.rds")$metadata

# Construct cell-type-specific gene expression signatures of SEs
avgexprs <- ComputeAvg(normdata = normdata, scmeta = metadata)
head(avgexprs)
#>                             SE09        SE01        SE08         SE07
#> Fibroblast..PDK4      0.35259620  0.13410953  0.06966571  0.118335435
#> Fibroblast..TNFRSF17  0.00000000  0.00000000  0.00000000  0.000000000
#> Fibroblast..ICAM3     0.00000000  0.00000000  0.00000000  0.000000000
#> Fibroblast..FAP      -0.29331695  0.02505667 -0.07346211  0.077473474
#> Fibroblast..GZMB     -0.09797678 -0.04104210 -0.09082236 -0.002063296
#> Fibroblast..TSC2     -0.05899541 -0.02658523 -0.02690466 -0.016824868
#>                             SE04        SE03         SE06       SE02       SE05
#> Fibroblast..PDK4     -0.04187951 -0.18236328 -0.121186218 0.02876511 0.01315714
#> Fibroblast..TNFRSF17  0.00000000  0.00000000  0.000000000 0.00000000 0.00000000
#> Fibroblast..ICAM3     0.00000000  0.00000000  0.000000000 0.00000000 0.00000000
#> Fibroblast..FAP       0.05147663 -0.08659207  0.053183806 0.02960319 0.11607862
#> Fibroblast..GZMB      0.01127651  0.02282926 -0.009071706 0.03888278 0.04906746
#> Fibroblast..TSC2      0.02419604 -0.04820711 -0.001317502 0.07683963 0.09783272