Compute Cell-Type-Specific Average Expression of Spatial Clusters

Compute Cell-Type-Specific Average Expression of Spatial Clusters

Usage

ComputeAvgs(
  normdata,
  scmeta,
  cluster = "SE",
  Region = NULL,
  scale = TRUE,
  ncores = 4
)

Arguments

normdata

Numeric matrix of normalized expression data, where rows represent genes and columns represent cells.

scmeta

Data frame containing metadata associated with each cell, including spatial cluster and cell type annotations.

cluster

Character string specifying the column name in 'scmeta' containing spatial cluster annotations.

Region

Character string specifying the column name in metadata data frames containing region annotations (default: NULL).

scale

A boolean specifying whether to do univariance normalization.

ncores

Integer specifying the number of CPU cores to use for parallel processing.

Value

A matrix of average expression, where rows represent genes and columns represent spatial clusters from the sample.

Examples

library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB"),
              "HumanMelanomaPatient1_subset_counts.tsv.gz", overwrite = TRUE)
#> File downloaded:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
#>   <id: 1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB>
#> Saved locally as:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
drive_download(as_id("1nSPj2zRywFUdbo1fwiz77ds4NuM6bmV2"),
              "Melanoma1_subset_SpatialEcoTyper_results.rds", overwrite = TRUE)
#> File downloaded:
#> • Melanoma1_subset_SpatialEcoTyper_results.rds
#>   <id: 1nSPj2zRywFUdbo1fwiz77ds4NuM6bmV2>
#> Saved locally as:
#> • Melanoma1_subset_SpatialEcoTyper_results.rds
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
tmpobj <- CreateSeuratObject(scdata) %>%
        SCTransform(clip.range = c(-10, 10), verbose = FALSE)
seurat_version = as.integer(gsub("\\..*", "", as.character(packageVersion("SeuratObject"))))
if(seurat_version<5){
  normdata <- GetAssayData(tmpobj, "data")
}else{
  normdata <- tmpobj[["SCT"]]$data
}
metadata = readRDS("Melanoma1_subset_SpatialEcoTyper_results.rds")$metadata

# Construct cell-type-specific gene expression signatures of SEs
avgexprs <- ComputeAvgs(normdata = normdata, scmeta = metadata)
head(avgexprs)
#>                              SE7          SE2          SE4          SE1
#> Fibroblast..PDK4      0.16877662  0.125425787  0.129077478 -0.048618850
#> Fibroblast..TNFRSF17  0.00000000  0.000000000  0.000000000  0.000000000
#> Fibroblast..ICAM3     0.00000000  0.000000000  0.000000000  0.000000000
#> Fibroblast..FAP      -0.16245753  0.008769572  0.076137301  0.051954130
#> Fibroblast..GZMB     -0.09869274 -0.027129039 -0.031853513 -0.007981858
#> Fibroblast..TSC2     -0.04840011 -0.037245579  0.009784921  0.053147770
#>                              SE6         SE0         SE3        SE5         SE8
#> Fibroblast..PDK4     -0.10288344 -0.18612179 -0.03040986 0.07024450  0.11917138
#> Fibroblast..TNFRSF17  0.00000000  0.00000000  0.00000000 0.00000000  0.00000000
#> Fibroblast..ICAM3     0.00000000  0.00000000  0.00000000 0.00000000  0.00000000
#> Fibroblast..FAP       0.04412527 -0.07947213  0.08198545 0.07579472 -0.02290463
#> Fibroblast..GZMB     -0.01636766  0.02581098  0.02334403 0.11797776  0.07086087
#> Fibroblast..TSC2     -0.04157964 -0.04773077  0.03064768 0.08075058  0.21924574