Integrate Multiple Spatial Transcriptomics Datasets to Identify Conserved Spatial Ecotypes

This function performs SpatialEcoTyper analysis on multiple spatial transcriptomics datasets. It normalizes the input data, performs SpatialEcoTyper analysis on each dataset, and integrates the results across samples.

Usage

MultiSpatialEcoTyper(
  data_list,
  metadata_list,
  outdir = "./",
  normalization.method = "None",
  nmf_ranks = 10,
  nrun.per.rank = 30,
  radius = 50,
  min.cts.per.region = 2,
  nfeatures = 3000,
  min.features = 10,
  Region = NULL,
  subresolution = 30,
  ncores = 1,
  seed = 1,
  ...
)

Arguments

data_list

A named list of expression matrices where each matrix represents gene expression data for a sample. The columns of each matrix correspond to cells, and the rows correspond to genes. Sample names should be used as list names. Otherwise, the samples will be named as 'Sample1' through 'SampleN'.

metadata_list

A named list of metadata data frames where each data frame contains metadata corresponding to the cells in the expression matrices. Each row should correspond to a column (cell) in the expression matrices. Each metadata should include at least three columns, including X, Y and CellType.

outdir

Directory where the results will be saved. Defaults to the current directory with a subdirectory named "SpatialEcoTyper_results_" followed by the current date.

normalization.method

Method for normalizing the expression data. Options include "None" (default), "SCT", or other methods compatible with Seurat's `NormalizeData` function.

nmf_ranks

Integer or a vector specifying the number of clusters (10 by default). When an integer vector is supplied, the function will test all supplied numbers and select the optimal number, which takes time.

nrun.per.rank

An integer specifying the the number of runs per rank for NMF (default: 30).

radius

Numeric specifying the radius (in the same units as spatial coordinates) for defining spatial neighborhoods around each cell. Default is 50.

min.cts.per.region

Integer specifying the minimum number of cell types (default: 2) required for a microregion.

nfeatures

An integer specifying the maximum number of top variable genes to select for each cell type.

min.features

An integer specifying the minimum number of shared features (genes) required across samples.

Region

Character string specifying the column name in metadata data frames containing region annotations (default: NULL). Pathologist annotation is recommended if available.

subresolution

Numeric specifying the resolution for clustering within each sample.

ncores

Integer specifying the number of cores for parallel processing. Default is 1.

seed

An integer used to seed the random number generator for NMF analysis.

...

Additional arguments passed to the `SpatialEcoTyper` function.

Value

The function saves the results in the specified output directory.

Details

This function takes a list of gene expression matrices and corresponding metadata, normalizes the data if specified, performs SpatialEcoTyper on each sample, and integrates the results across multiple samples to identify conserved spatial ecotypes.

Examples

# See https://digitalcytometry.github.io/SpatialEcoTyper/docs/articles/Integration.html