Compute Cell-Type-Specific Average Expression of Spatial Clusters

Compute Cell-Type-Specific Average Expression of Spatial Clusters

Usage

ComputeAvgs(
  normdata,
  scmeta,
  cluster = "SE",
  Region = NULL,
  scale = TRUE,
  ncores = 4
)

Arguments

normdata

Numeric matrix of normalized expression data, where rows represent genes and columns represent cells.

scmeta

Data frame containing metadata associated with each cell, including spatial cluster and cell type annotations.

cluster

Character string specifying the column name in 'scmeta' containing spatial cluster annotations.

Region

Character string specifying the column name in metadata data frames containing region annotations (default: NULL).

scale

A boolean specifying whether to do univariance normalization.

ncores

Integer specifying the number of CPU cores to use for parallel processing.

Value

A matrix of average expression, where rows represent genes and columns represent spatial clusters from the sample.

Examples

library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB"),
              "HumanMelanomaPatient1_subset_counts.tsv.gz", overwrite = TRUE)
#> File downloaded:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
#>   <id: 1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB>
#> Saved locally as:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
drive_download(as_id("1nSPj2zRywFUdbo1fwiz77ds4NuM6bmV2"),
              "Melanoma1_subset_SpatialEcoTyper_results.rds", overwrite = TRUE)
#> Error in map(as_id(id), get_one_file_id): ℹ In index: 1.
#> Caused by error in `.f()`:
#> ! Client error: (404) Not Found
#> File not found: 1nSPj2zRywFUdbo1fwiz77ds4NuM6bmV2.
#> • message: File not found: 1nSPj2zRywFUdbo1fwiz77ds4NuM6bmV2.
#> • domain: global
#> • reason: notFound
#> • location: fileId
#> • locationType: parameter
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
tmpobj <- CreateSeuratObject(scdata) %>%
        SCTransform(clip.range = c(-10, 10), verbose = FALSE)
#> Warning: Data is of class matrix. Coercing to dgCMatrix.
seurat_version = as.integer(gsub("\\..*", "", as.character(packageVersion("SeuratObject"))))
if(seurat_version<5){
  normdata <- GetAssayData(tmpobj, "data")
}else{
  normdata <- tmpobj[["SCT"]]$data
}
metadata = readRDS("Melanoma1_subset_SpatialEcoTyper_results.rds")$metadata

# Construct cell-type-specific gene expression signatures of SEs
avgexprs <- ComputeAvgs(normdata = normdata, scmeta = metadata)
#> Error in ComputeAvgs(normdata = normdata, scmeta = metadata): could not find function "ComputeAvgs"
head(avgexprs)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'avgexprs' not found