Construct Spatial Metacells from Single-Cell Spatial Data

This function computes spatial metacells from single-cell spatial transcriptomics. Spatial metacells represent aggregated expression profiles for spatially proximal cells. The function utilizes k-nearest neighbor (KNN) weighting to aggregate expression profiles from neighboring cells and generate metacell expression profiles for each cell type.

Usage

GetSpatialMetacells(
  normdata,
  metadata,
  X = "X",
  Y = "Y",
  CellType = "CellType",
  k = 20,
  radius = 50,
  bin = "SpotID",
  bin.X = "Spot.X",
  bin.Y = "Spot.Y",
  min.cells.per.region = 2,
  ncores = 1
)

Arguments

normdata

A matrix or data frame representing normalized gene expression data, where rows correspond to genes and columns correspond to cells.

metadata

A data frame containing metadata associated with each cell. Must include spatial coordinates (e.g., X and Y) as well as cell type annotations. The row names of the metadata must match the column names of the normdata.

X

Character string specifying the column name in metadata containing the X spatial coordinates.

Y

Character string specifying the column name in metadata containing the Y spatial coordinates.

CellType

Character string specifying the column name in metadata containing the cell type annotations.

k

Integer specifying the number of nearest spatial neighbors for constructing metacells.

radius

Numeric value specifying the radius (in units of spatial coordinates) within which neighboring cells are considered.

bin

Character string specifying the column name in metadata containing the microregion identifier.

bin.X

Character string specifying the column name in metadata containing the X spatial coordinates of microregions.

bin.Y

Character string specifying the column name in metadata containing the Y spatial coordinates of microregions.

min.cells.per.region

Integer specifying the minimum number of cells required in each spatial microregion to compute metacells.

ncores

Integer specifying the number of CPU cores to use for parallel processing (default: 1).

Value

A matrix containing the spatial metacell expression profiles, with rows representing individual genes and columns representing metacells.

Examples

library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("1CgUOQKrWY_TG61o5aw7J9LZzE20D6NuI"),
                    "HumanMelanomaPatient1_subset_scmeta.tsv", overwrite = TRUE)
#> File downloaded:
#> • HumanMelanomaPatient1_subset_scmeta.tsv
#>   <id: 1CgUOQKrWY_TG61o5aw7J9LZzE20D6NuI>
#> Saved locally as:
#> • HumanMelanomaPatient1_subset_scmeta.tsv
drive_download(as_id("1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB"),
              "HumanMelanomaPatient1_subset_counts.tsv.gz", overwrite = TRUE)
#> File downloaded:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
#>   <id: 1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB>
#> Saved locally as:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
normdata = NormalizeData(scdata)
scmeta <- read.table("HumanMelanomaPatient1_subset_scmeta.tsv",
                      sep = "\t",header = TRUE, row.names = 1)

# Construct spatial metacells from single-cell spatial data
metacells <- GetSpatialMetacells(normdata = normdata, metadata = scmeta,
                                 X = "X", Y = "Y", CellType = "CellType",
                                 min.cells.per.region = 2)
head(metacells)
#> 6 x 390 sparse Matrix of class "dgCMatrix"
#>   [[ suppressing 390 column names ‘X-91_Y-91..B’, ‘X-92_Y-92..B’, ‘X-93_Y-93..B’ ... ]]
#>                                                                             
#> PDK4     .         .         .         .         .        .         .       
#> TNFRSF17 0.6544829 .         0.4175847 .         1.264464 2.0349031 .       
#> ICAM3    2.4550440 2.7259684 2.6240964 3.3432630 4.074004 1.5776727 2.339249
#> FAP      .         .         .         .         .        .         .       
#> GZMB     .         0.2362506 0.7086796 .         .        0.5231711 .       
#> TSC2     .         .         .         0.5619819 .        .         .       
#>                                                                           
#> PDK4     .        .         .         .        .        .        .        
#> TNFRSF17 2.075125 0.6507723 0.7586164 0.803498 1.063637 0.293977 0.6992404
#> ICAM3    2.819178 2.4270566 3.0788990 1.808787 3.170151 1.974216 2.5348550
#> FAP      .        .         .         .        .        .        .        
#> GZMB     .        0.4860334 .         .        .        .        0.2107297
#> TSC2     .        .         .         .        .        .        .        
#>                                                                              
#> PDK4     .         .         .         .        .         .         .        
#> TNFRSF17 0.4269010 0.7234476 0.4021635 .        0.2771736 0.3897001 0.8802495
#> ICAM3    2.5542980 2.9837018 3.4999791 3.140392 2.1379927 2.6634697 1.8515150
#> FAP      .         .         .         .        .         .         .        
#> GZMB     0.4153259 0.6800209 0.2671964 .        0.6743574 .         .        
#> TSC2     .         .         0.2283413 .        .         .         .        
#>                                                                               
#> PDK4     .         .         .         .         .         .         .        
#> TNFRSF17 0.6444861 1.1535714 0.2124001 0.2698584 0.2260400 0.7285162 0.6506251
#> ICAM3    2.8541354 2.6135363 3.0868468 2.8681769 2.9581242 2.5694730 2.8433845
#> FAP      .         .         .         .         .         .         .        
#> GZMB     .         0.6531461 .         0.4481639 0.2221523 .         .        
#> TSC2     .         .         .         0.2124001 0.2130847 .         .        
#>                                                                             
#> PDK4     .         .         .         .        .         .         .       
#> TNFRSF17 0.2081711 0.2327773 0.3701375 .        0.9364823 0.8657987 0.840786
#> ICAM3    2.0815837 2.9877662 3.5032758 2.231905 2.9240027 3.2251307 1.863237
#> FAP      .         .         .         .        .         .         .       
#> GZMB     0.2399883 0.4576690 0.6595242 .        .         0.1949055 .       
#> TSC2     0.2209223 .         .         .        .         0.1991069 .       
#>                                                                               
#> PDK4     .        .         .         .        .         . .         .        
#> TNFRSF17 1.210397 0.6245759 0.4394449 .        0.9210538 . .         0.5213488
#> ICAM3    2.073252 2.6591540 3.2826420 2.129434 1.6783614 . 0.8773159 2.8655548
#> FAP      .        .         .         .        .         . .         .        
#> GZMB     .        0.3124665 0.6175867 0.452080 0.7573076 . 0.8339345 .        
#> TSC2     .        .         .         .        .         . .         .        
#>                                                                           
#> PDK4     .         .         .        .         .       .         .       
#> TNFRSF17 0.9123604 .         .        .         .       .         .       
#> ICAM3    3.3184522 3.2481927 2.661054 2.1067963 2.98845 2.6580686 2.694117
#> FAP      .         .         .        .         .       .         .       
#> GZMB     .         0.2159238 0.207552 0.2622667 .       .         .       
#> TSC2     .         0.2343499 .        .         .       0.2225693 .       
#>                                                                           
#> PDK4     .        .         .        .        .        .         .        
#> TNFRSF17 .        .         .        .        .        .         .        
#> ICAM3    2.352396 1.6403401 2.153412 3.261016 1.642338 2.6158517 4.0367157
#> FAP      .        .         .        .        .        .         .        
#> GZMB     .        0.2513632 .        .        .        .         0.1935062
#> TSC2     .        .         .        .        .        0.4806539 .        
#>                                                                         
#> PDK4     .        .         .        .        .        .        .       
#> TNFRSF17 .        .         .        .        .        .        .       
#> ICAM3    3.718786 2.6932140 3.384322 1.734639 1.850746 3.247609 2.267498
#> FAP      .        .         .        .        .        .        .       
#> GZMB     .        0.1961036 .        .        .        .        0.275043
#> TSC2     .        .         .        .        .        .        .       
#>                                                                             
#> PDK4     .        .        .         .         .         .         .        
#> TNFRSF17 .        .        .         .         .         .         .        
#> ICAM3    2.209121 3.661897 2.9055060 2.4884698 3.0916244 3.3684981 2.1944636
#> FAP      .        .        .         .         .         .         .        
#> GZMB     .        .        0.4866689 0.2450162 .         0.1988493 0.2411979
#> TSC2     .        .        0.2014891 .         0.2017612 .         .        
#>                                                                              
#> PDK4     .         .         .         .         .         .        .        
#> TNFRSF17 .         .         .         .         .         .        .        
#> ICAM3    2.8888087 2.3907300 2.5080751 1.8866217 2.3343320 2.615067 2.0112396
#> FAP      .         .         .         .         .         .        .        
#> GZMB     0.6999925 1.3139516 0.2651652 0.2081711 0.2709646 .        0.2057408
#> TSC2     0.2031452 0.2274077 0.4654402 0.2661704 .         .        .        
#>                                                                            
#> PDK4     .         .         .        .         .        .         .       
#> TNFRSF17 .         .         .        .         .        .         .       
#> ICAM3    3.0613972 2.4068099 2.447891 3.2194577 1.938482 2.6015475 2.700411
#> FAP      .         .         .        .         .        .         .       
#> GZMB     .         0.5289607 .        0.2687764 .        .         .       
#> TSC2     0.2463577 .         .        .         .        0.4641257 .       
#>                                                                             
#> PDK4     .        .         .         .         .         .        .        
#> TNFRSF17 .        .         .         .         .         .        .        
#> ICAM3    2.269408 0.9489672 2.5817088 1.8017008 1.9775697 2.605423 2.8804254
#> FAP      .        .         .         .         .         .        .        
#> GZMB     .        .         0.4314681 1.6397395 0.5287651 1.193708 1.4091318
#> TSC2     .        .         .         0.2393944 0.2213289 .        0.2051518
#>                                                                              
#> PDK4     .         .         .         .         .         .        .        
#> TNFRSF17 .         .         .         .         .         .        .        
#> ICAM3    2.6076950 3.3470302 2.2187896 2.7504711 2.6651482 2.108225 2.7845421
#> FAP      .         .         .         .         .         .        .        
#> GZMB     0.8810538 0.9619033 0.7208446 1.2240569 0.7219464 1.092021 0.5355330
#> TSC2     .         0.2560989 .         0.2370883 .         .        0.4995151
#>                                                                           
#> PDK4     .         .        .        .         .        .        .        
#> TNFRSF17 .         .        .        .         .        .        .        
#> ICAM3    1.0805223 2.539781 1.418539 2.7131510 2.510669 2.304236 2.1329565
#> FAP      .         .        .        .         .        .        .        
#> GZMB     1.7279168 1.424386 1.128606 0.2677175 1.204423 1.506020 0.9323715
#> TSC2     0.1662692 .        .        0.1758448 .        .        .        
#>                                                                          
#> PDK4     .        .        .         .         .        .        .       
#> TNFRSF17 .        .        .         .         .        .        .       
#> ICAM3    3.326815 3.132616 2.3915625 2.5171377 3.042002 2.757046 2.939314
#> FAP      .        .        .         .         .        .        .       
#> GZMB     1.947208 1.595812 1.8038635 0.4736165 1.880746 1.861779 1.196432
#> TSC2     .        .        0.2544693 0.1951427 0.248441 .        .       
#>                                                                             
#> PDK4     .         .         .         .        .        .         .        
#> TNFRSF17 .         .         .         .        .        .         .        
#> ICAM3    2.0389190 1.8700239 2.6950142 3.025181 2.951156 1.0938623 1.8833771
#> FAP      .         .         .         .        .        .         .        
#> GZMB     0.5160454 0.5196073 0.8921513 1.410914 1.403760 1.0841155 0.6744100
#> TSC2     0.1749030 .         0.2288148 .        .        0.4759744 0.2327773
#>                                                                             
#> PDK4     .         .         .        .        .         .         .        
#> TNFRSF17 .         .         .        .        .         .         .        
#> ICAM3    1.5948809 1.4559048 2.797505 1.669694 2.9915786 3.0622097 2.0882418
#> FAP      .         .         .        .        .         .         .        
#> GZMB     0.4486093 0.4655083 1.112943 1.134259 1.3371663 0.8640992 0.7224095
#> TSC2     0.1817480 0.2278723 0.191481 .        0.4202619 .         .        
#>                                                                              
#> PDK4     .         .         .        .         .         0.6721268 0.5777224
#> TNFRSF17 .         .         .        .         .         .         .        
#> ICAM3    2.6304309 2.8175271 1.799939 1.8560540 2.8720481 2.0909612 1.4339229
#> FAP      .         .         .        .         .         .         .        
#> GZMB     0.9174394 2.1903889 1.356037 0.7205789 0.9291368 .         .        
#> TSC2     0.2569343 0.2017612 .        0.2247095 0.2552774 0.4161715 .        
#>                                                                              
#> PDK4     0.2151848 1.9538824 1.550975 0.2593085 0.8897096 0.5263809 0.9804341
#> TNFRSF17 .         .         .        .         .         .         .        
#> ICAM3    0.8862776 1.0964768 2.350320 1.6117519 1.4258686 1.6114262 2.2472793
#> FAP      .         .         .        .         .         .         .        
#> GZMB     0.1918842 .         .        .         0.3659154 .         0.3811077
#> TSC2     .         0.1694787 .        .         0.2812897 .         0.1749030
#>                                                                             
#> PDK4     1.267374 0.7523609 1.410662 0.6016010 0.6998461 1.2510781 0.7752687
#> TNFRSF17 .        .         .        .         .         .         .        
#> ICAM3    1.684209 2.8418969 1.805333 2.1237841 0.2838192 2.8558944 1.5863079
#> FAP      .        .         .        .         .         .         .        
#> GZMB     0.252164 .         .        .         .         0.3884015 .        
#> TSC2     .        .         .        0.3232802 .         .         0.1637539
#>                                                                           
#> PDK4     1.344222 0.8693882 1.962452 0.7557593 0.9761035 .        1.302667
#> TNFRSF17 .        .         .        .         .         .        .       
#> ICAM3    1.421307 1.2766308 2.510840 1.2947254 1.3729438 4.205427 2.078886
#> FAP      .        .         .        .         .         .        .       
#> GZMB     .        .         .        .         .         .        .       
#> TSC2     .        .         .        .         .         .        .       
#>                                                                          
#> PDK4     1.3642321 1.0150225 .       0.691606 1.110163 1.077798 .        
#> TNFRSF17 .         .         .       .        .        .        .        
#> ICAM3    2.2525109 0.4720003 2.95211 1.404622 2.273123 1.077798 0.5165948
#> FAP      .         .         .       .        .        .        .        
#> GZMB     .         .         .       .        .        .        .        
#> TSC2     0.7564669 .         .       .        .        .        .        
#>                                                                           
#> PDK4     1.308167 0.560643 1.087698 0.7743722 0.7405078 0.9220588 .       
#> TNFRSF17 .        .        .        .         .         .         .       
#> ICAM3    2.555106 1.551848 1.030582 0.3995826 2.2734382 1.5979620 2.174079
#> FAP      .        .        .        .         .         .         .       
#> GZMB     .        .        .        .         .         0.4333204 .       
#> TSC2     .        .        .        .         .         .         .       
#>                                                                            
#> PDK4     2.002619 0.5654545 0.5092751 1.405295 0.5031320 0.825989 1.2666457
#> TNFRSF17 .        .         .         .        .         .        .        
#> ICAM3    1.181576 1.9442164 0.8655169 1.415085 1.8753491 2.381560 1.4720859
#> FAP      .        .         .         .        .         .        .        
#> GZMB     .        0.6413099 .         .        0.5180635 .        .        
#> TSC2     .        0.2633816 .         .        .         .        0.2417005
#>                                                                              
#> PDK4     1.1627010 0.1886991 0.7742990 0.6019194 0.6623059 0.5476020 1.213516
#> TNFRSF17 .         .         .         .         .         .         .       
#> ICAM3    .         .         .         .         .         .         .       
#> FAP      0.3893810 0.4059865 .         0.1841427 0.1424201 0.1626651 .       
#> GZMB     0.1694787 .         0.1745932 .         0.3653346 .         .       
#> TSC2     .         .         .         0.2120614 0.1884917 .         .       
#>                                                                               
#> PDK4     0.7018900 1.1720950 1.0525231 0.7075257 0.3193786 0.6871473 0.6254852
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      .         0.7419234 0.8879133 0.1613681 0.3852947 0.1682491 0.1714580
#> GZMB     .         0.2463577 .         .         0.7868361 .         0.1939681
#> TSC2     0.1597685 .         .         0.3196826 .         0.1682491 .        
#>                                                                               
#> PDK4     0.2484410 0.1530914 1.2657517 0.7845304 1.1712250 0.8447723 0.9651236
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      0.2134307 0.2368074 0.1847129 0.9131424 .         .         0.8272398
#> GZMB     .         0.2463577 0.2084836 0.1476121 0.1985932 0.1229100 0.4513565
#> TSC2     .         .         .         .         .         .         0.2260400
#>                                                                               
#> PDK4     0.6396284 0.9225859 1.2567108 0.9416234 0.6257305 0.6183192 0.2459065
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      .         0.3777682 0.3482939 0.3488451 .         0.5958435 0.1998880
#> GZMB     0.4516062 .         0.1625456 .         0.1572555 .         .        
#> TSC2     .         .         0.1455939 .         .         .         .        
#>                                                                               
#> PDK4     1.1436099 0.1605614 0.7291590 0.3314464 0.9711474 0.2317563 0.4959660
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      0.5111735 .         0.4105836 0.4014640 0.3871480 0.4421309 0.3712461
#> GZMB     .         .         .         0.4216995 .         0.4316623 .        
#> TSC2     0.2066379 .         .         .         0.1824721 .         .        
#>                                                                               
#> PDK4     0.9550457 0.2069407 0.9510292 1.7775318 1.0277420 0.3306178 1.1001530
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      0.1849044 .         0.3555773 0.2025868 0.1665283 0.3937549 0.1341966
#> GZMB     .         .         0.1889075 .         .         .         0.1733741
#> TSC2     0.3446825 .         .         0.2185514 .         .         .        
#>                                                                               
#> PDK4     0.9136417 0.7375121 0.8442152 0.3508610 0.2307560 0.5853667 0.1928216
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      .         0.3796194 1.7051766 0.5534571 0.9086273 0.2909049 0.3478969
#> GZMB     .         0.8436475 0.1897500 .         .         .         .        
#> TSC2     .         .         .         .         .         .         .        
#>                                                                               
#> PDK4     0.3833703 0.5800734 0.3820236 0.6046471 0.6096568 0.4500156 0.1479574
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      1.1188588 1.1273693 0.6267170 0.3553790 0.3580661 0.8820832 1.5345110
#> GZMB     .         .         .         .         0.1382560 0.4831283 0.2117250
#> TSC2     0.1258949 .         .         .         .         .         .        
#>                                                                              
#> PDK4     0.2042816 0.2063370 1.1307577 .        0.7146263 1.1447852 .        
#> TNFRSF17 .         .         .         .        .         .         .        
#> ICAM3    .         .         .         .        .         .         .        
#> FAP      0.7962302 1.2130316 0.5017743 0.510985 1.1977907 0.9339003 0.4954266
#> GZMB     .         0.1903918 0.1573619 0.191481 0.1689282 .         .        
#> TSC2     .         .         .         .        .         .         0.2148396
#>                                                                               
#> PDK4     0.1517524 0.5670232 0.8698162 0.1637539 0.3696494 0.1016129 0.7851283
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      1.1275889 0.5638223 0.4307200 0.3936369 0.6520385 0.9392034 0.6904686
#> GZMB     .         0.4052187 0.3526905 .         .         .         .        
#> TSC2     .         .         .         .         0.1944345 .         .        
#>                                                                               
#> PDK4     0.2054454 0.3601386 0.3271071 0.7331518 0.4034587 0.5518261 0.1769679
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      0.7812682 0.9995687 0.7188326 0.7873114 0.1610207 0.1876709 0.4363381
#> GZMB     .         0.1641226 0.4060103 0.1396010 .         .         0.1696173
#> TSC2     .         0.3449809 0.4130268 0.1998880 0.1723327 .         .        
#>                                                                             
#> PDK4     0.6366271 0.5125880 0.9258237 0.2558235 .        0.430857 0.3606961
#> TNFRSF17 .         .         .         .         .        .        .        
#> ICAM3    .         .         .         .         .        .        .        
#> FAP      0.5623080 0.9737224 0.9275480 0.3332545 1.174189 0.189963 0.4993355
#> GZMB     .         0.1771306 .         .         .        .        .        
#> TSC2     .         .         .         0.1886991 .        .        .        
#>                                                                               
#> PDK4     0.3619712 0.3824599 0.3979787 0.5289367 1.0244530 1.4071443 1.2033558
#> TNFRSF17 .         .         .         .         .         .         .        
#> ICAM3    .         .         .         .         .         .         .        
#> FAP      0.5294469 0.3884222 0.9865208 0.8786052 .         .         .        
#> GZMB     0.4239187 0.1761630 .         .         0.1891167 0.1925955 0.4594442
#> TSC2     0.3623193 .         .         .         .         .         .        
#>                                                                              
#> PDK4     0.9370442 0.4295889 1.649515 0.4096543 0.6272938 0.6509813 1.1178334
#> TNFRSF17 .         .         .        .         .         .         .        
#> ICAM3    .         .         .        .         .         .         .        
#> FAP      .         .         .        .         .         .         .        
#> GZMB     .         0.3762156 .        0.2443589 0.1843320 .         0.2292929
#> TSC2     .         .         .        0.2443589 .         0.2104024 0.1739797
#>                                                                              
#> PDK4     1.0781587 0.6916388 0.6727012 1.328545 0.8761082 0.8068752 0.8351528
#> TNFRSF17 .         .         .         .        .         .         .        
#> ICAM3    .         .         .         .        .         .         .        
#> FAP      .         .         .         .        .         .         .        
#> GZMB     0.4975071 0.2213289 .         .        0.2480876 .         .        
#> TSC2     .         .         .         .        .         0.1805070 .        
#>                                                                             
#> PDK4     1.685659 0.6121596 1.3239812 0.3969014 1.604153 1.3238941 0.8400419
#> TNFRSF17 .        .         .         .         .        .         .        
#> ICAM3    .        .         .         .         .        .         .        
#> FAP      .        .         .         .         .        .         .        
#> GZMB     .        0.2569343 0.4008008 0.2638560 .        0.1854839 0.4082143
#> TSC2     .        .         .         0.1874679 .        0.1895380 0.1657551
#>                                                                              
#> PDK4     0.3858890 1.0302486 0.7097701 1.0220129 1.492092 0.7326239 0.9761929
#> TNFRSF17 .         .         .         .         .        .         .        
#> ICAM3    .         .         .         .         .        .         .        
#> FAP      .         .         .         .         .        .         .        
#> GZMB     0.1638765 0.2002833 .         0.1485682 .        0.3916174 0.3835622
#> TSC2     .         0.3567406 .         .         .        .         0.4402635
#>                                                                              
#> PDK4     1.6147421 0.8084328 2.0180415 0.4060164 1.075883 0.5614577 1.2615632
#> TNFRSF17 .         .         .         .         .        .         .        
#> ICAM3    .         .         .         .         .        .         .        
#> FAP      .         .         .         .         .        .         .        
#> GZMB     0.1542714 0.2075520 .         .         .        .         0.5762879
#> TSC2     0.1736759 0.1638765 0.2017612 .         .        0.1249105 .        
#>                                                                              
#> PDK4     0.6241975 0.5578205 0.9675281 0.6808135 0.4154361 0.5436584 .       
#> TNFRSF17 .         .         .         .         .         .         .       
#> ICAM3    .         .         .         .         .         .         1.223828
#> FAP      .         .         .         .         .         .         .       
#> GZMB     0.1946694 0.1874679 .         .         .         0.5046447 2.072505
#> TSC2     0.1669198 0.1956208 .         .         .         0.1644943 .       
#>                                                                             
#> PDK4     .        .        .         .         .         .         .        
#> TNFRSF17 .        .        .         .         .         .         .        
#> ICAM3    1.802485 2.075318 3.2894150 0.9697523 0.8145297 2.8004099 2.3900841
#> FAP      .        .        .         .         .         .         .        
#> GZMB     .        1.329268 0.5548593 .         0.8469007 .         0.6125405
#> TSC2     .        .        .         .         .         0.5645194 1.0878796
#>                                                                            
#> PDK4     .        .         .         . .        .        .        .       
#> TNFRSF17 .        .         .         . .        .        .        .       
#> ICAM3    2.952899 3.1949526 1.9814483 . 1.983549 2.934933 2.334114 .       
#> FAP      .        .         .         . .        .        .        .       
#> GZMB     2.880698 1.9992169 0.9343665 . 1.545959 4.111746 .        .       
#> TSC2     1.558801 0.9937641 .         . .        .        .        1.670775
#>                                                                         
#> PDK4     .        .        .         .        .        .        .       
#> TNFRSF17 .        .        .         .        .        .        .       
#> ICAM3    3.117911 4.945499 2.6751936 2.361608 4.944884 2.432021 3.366057
#> FAP      .        .        .         .        .        .        .       
#> GZMB     3.153962 .        1.5653551 1.162464 1.748497 1.209149 3.500917
#> TSC2     .        .        0.4839691 .        .        .        .       
#>                                                                         
#> PDK4     .        .        .        .         .        .        .       
#> TNFRSF17 .        .        .        .         .        .        .       
#> ICAM3    1.151317 4.126723 4.435517 2.1636765 .        2.581101 3.896530
#> FAP      .        .        .        .         .        .        .       
#> GZMB     3.749509 .        .        1.4956185 .        .        3.066501
#> TSC2     1.245778 .        .        0.4798306 1.253081 .        .       
#>                                                                        
#> PDK4     .        .        .        .        .        .        .       
#> TNFRSF17 .        .        .        .        .        .        .       
#> ICAM3    2.608787 1.185441 1.639332 1.569513 3.021302 3.206785 0.871931
#> FAP      .        .        .        .        .        .        .       
#> GZMB     .        2.113686 1.073360 1.535090 1.258004 2.979279 1.360329
#> TSC2     .        .        .        .        .        .        .       
#>                                                                           
#> PDK4     .        .        0.6972153 0.1884917 .        0.7939925 .       
#> TNFRSF17 .        .        1.2165561 1.9935838 1.183298 0.7751313 2.016898
#> ICAM3    2.269476 3.220389 2.6474173 3.1150602 3.457400 1.3780484 2.127905
#> FAP      .        .        .         .         .        .         .       
#> GZMB     .        .        1.5317956 1.5563737 1.853594 3.0576691 1.999913
#> TSC2     .        .        .         .         .        .         .       
#>                                                                             
#> PDK4     0.7364076 .        0.5685193 0.9434252 .         0.3331466 1.758241
#> TNFRSF17 1.4805428 1.166117 2.3278663 1.4651472 0.6402209 1.5421945 3.007102
#> ICAM3    1.5016807 2.900115 2.5482026 2.1274349 1.6713046 2.2073857 1.886853
#> FAP      .         .        .         .         .         .         .       
#> GZMB     4.5555136 1.467949 1.4404367 3.1550420 1.3500698 1.8767081 1.283244
#> TSC2     .         .        .         .         .         .         .       
#>                                                                           
#> PDK4     0.5477196 0.7675449 0.681212 1.239706 1.3878340 1.435087 .       
#> TNFRSF17 1.2680576 0.7222006 2.793247 1.279220 0.6486377 1.115905 2.739642
#> ICAM3    2.7000409 2.2182937 1.353881 2.961550 3.0513625 2.516795 2.101882
#> FAP      .         .         .        .        .         .        .       
#> GZMB     1.7595718 1.9452350 1.516360 2.850711 0.3449012 1.884570 2.386950
#> TSC2     .         .         .        .        .         .        .       
#>                                                                             
#> PDK4     1.446884 0.7552811 0.5931933 0.2373661 0.5956401 0.4417934 .       
#> TNFRSF17 2.826489 1.7379933 0.9365347 0.5284846 1.2979420 1.1825275 .       
#> ICAM3    2.749241 3.2720699 3.2887507 3.3479078 2.9612091 3.0282056 3.693063
#> FAP      .        .         .         .         .         .         .       
#> GZMB     .        1.3864504 1.7920635 1.4293033 1.2834571 0.5312479 1.096959
#> TSC2     .        .         .         .         .         .         .       
#>                                                                             
#> PDK4     .        0.6114597 0.3996979 0.9234839 0.2174079 1.068628 .        
#> TNFRSF17 1.628742 1.2010023 1.8421455 1.2845608 0.9884584 1.498147 0.2455236
#> ICAM3    2.925930 3.4709615 2.9111497 3.7205964 3.0424221 2.789474 2.8523507
#> FAP      .        .         .         .         .         .        .        
#> GZMB     1.192256 1.0596876 1.1691955 1.2390702 0.4948967 1.285955 1.0761320
#> TSC2     .        .         .         .         .         .        .        
#>                                                                             
#> PDK4     .        0.4598034 .         .         1.399232 0.2059514 0.8081269
#> TNFRSF17 4.498930 0.4113915 0.9921520 0.6034497 .        1.3432743 0.7137668
#> ICAM3    4.358957 2.6201706 0.5793906 2.4784173 1.816917 2.3051774 1.8183247
#> FAP      .        .         .         .         .        .         .        
#> GZMB     .        3.5560720 1.6287532 5.0245933 2.325537 1.9650682 4.3522638
#> TSC2     .        .         .         .         .        .         .        
#>                                     
#> PDK4     1.140120 .        0.1945096
#> TNFRSF17 1.386295 1.689997 1.5427200
#> ICAM3    3.471813 1.762560 3.3795050
#> FAP      .        .        .        
#> GZMB     2.013692 2.009341 1.6912014
#> TSC2     .        .        .