Identify Spatial EcoTypes from Single-cell Spatial Data (A Single Sample)

This function identifies spatially distinct cellular ecosystems (SE) from a single sample.

Usage

SpatialEcoTyper(
  normdata,
  metadata,
  outprefix = "SE",
  radius = 50,
  resolution = 0.8,
  nfeatures = 3000,
  npcs = 20,
  k.sn = 50,
  k = 20,
  min.cells = 5,
  min.features = 10,
  min.cts.per.region = 2,
  min.cells.per.region = 2,
  iterations = 5,
  ncores = 1,
  binsize = round(radius * 1.4)
)

Arguments

normdata

A matrix representing normalized gene expression data, where rows correspond to genes and columns correspond to cells.

metadata

A data frame containing metadata associated with each cell. Must include spatial coordinates (e.g., X and Y) as well as cell type annotations. The row names of the metadata must match the column names of the normdata.

outprefix

Character string specifying the prefix for output file names.

radius

Numeric specifying the radius (in the same units as spatial coordinates) for defining spatial neighborhoods around each cell. Default is 50.

resolution

Numeric specifying the resolution for Louvain clustering (default: 0.8).

nfeatures

Integer specifying the number of top variable features (default: 3000) used for PCA.

npcs

Integer specifying the number of principal components (PCs) (default: 20).

k.sn

Integer specifying the number of spatial nearest neighbors (default: 50) for constructing similarity network.

k

Integer specifying the number of spatial nearest neighbors (default: 20) used to construct spatial meta-cells.

min.cells

Minimum number of cells / spatial-meta-cells (default: 5) expressing a feature/gene.

min.features

Minimum number of features (default: 10) detected in a cell / spatial-meta-cell.

min.cts.per.region

Integer specifying the minimum number of cell types (default: 2) required for a microregion.

min.cells.per.region

Integer specifying the minimum number of cells (default: 2) in each spatial microregion.

iterations

Integer specifying the number of iterations (default: 5) for SNF analysis.

ncores

Integer specifying the number of CPU cores to use for parallel processing.

binsize

Numeric specifying the bin size for spatial discretization of coordinates. By default, this size is determined based on the specified radius (radius*1.4 µm). Increasing the binsize will downsample microregions and expedite the analysis, while it might eliminate cells located between bins from the SE discovery analysis.

#' @return A list containing two elements:

obj

A seurat object constructed from fused similarity network of sptial microregions

metadata

Updated metadata, with a new column (`SE`) added

Examples

# See https://digitalcytometry.github.io/SpatialEcoTyper/docs/articles/SingleSample.html
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(ggplot2))
suppressPackageStartupMessages(library(parallel))
suppressPackageStartupMessages(library(Seurat))
suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(googledrive))
suppressPackageStartupMessages(library(R.utils))
library(SpatialEcoTyper)

drive_deauth() # Disable Google sign-in requirement
drive_download(as_id("13Rc5Rsu8jbnEYYfUse-xQ7ges51LcI7n"), "HumanMelanomaPatient1_subset_counts.tsv.gz")
#> Error in drive_download(as_id("13Rc5Rsu8jbnEYYfUse-xQ7ges51LcI7n"), "HumanMelanomaPatient1_subset_counts.tsv.gz"): Local `path` already exists and overwrite is `FALSE`:
#> • HumanMelanomaPatient1_subset_counts.tsv.gz
drive_download(as_id("12xcZNhpT-xbhcG8kX1QAdTeM9TKeFAUW"), "HumanMelanomaPatient1_subset_scmeta.tsv")
#> Error in drive_download(as_id("12xcZNhpT-xbhcG8kX1QAdTeM9TKeFAUW"), "HumanMelanomaPatient1_subset_scmeta.tsv"): Local `path` already exists and overwrite is `FALSE`:
#> • HumanMelanomaPatient1_subset_scmeta.tsv

# Load single-cell gene expression matrix. Rows are genes, columns are cells.
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]  # set genes as row names
scdata <- as.matrix(scdata[, -1])
normdata <- NormalizeData(scdata)
head(normdata[, 1:5])
#> 6 x 5 sparse Matrix of class "dgCMatrix"
#>          HumanMelanomaPatient1__cell_3655 HumanMelanomaPatient1__cell_3657
#> PDK4                             .                                4.221184
#> TNFRSF17                         .                                .       
#> ICAM3                            .                                .       
#> FAP                              3.552438                         .       
#> GZMB                             .                                .       
#> TSC2                             .                                .       
#>          HumanMelanomaPatient1__cell_3658 HumanMelanomaPatient1__cell_3660
#> PDK4                             3.208945                                .
#> TNFRSF17                         .                                       .
#> ICAM3                            .                                       .
#> FAP                              .                                       .
#> GZMB                             .                                       .
#> TSC2                             .                                       .
#>          HumanMelanomaPatient1__cell_3661
#> PDK4                                    .
#> TNFRSF17                                .
#> ICAM3                                   .
#> FAP                                     .
#> GZMB                                    .
#> TSC2                                    .

# Load single-cell metadata. Three columns are required, including X, Y, and CellType. Others are optional.
scmeta <- read.table("HumanMelanomaPatient1_subset_scmeta.tsv",
                     sep = "\t",header = TRUE, row.names = 1)
scmeta <- scmeta[colnames(scdata), ] # match the cell ids in scdata and scmeta
head(scmeta)
#>                                         X         Y   CellType CellTypeName
#> HumanMelanomaPatient1__cell_3655 1894.706 -6367.766 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_3657 1942.480 -6369.602 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_3658 1963.007 -6374.026 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_3660 1981.600 -6372.266 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_3661 1742.939 -6374.851 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_3663 1921.683 -6383.309 Fibroblast  Fibroblasts
#>                                  Region Dist2Interface
#> HumanMelanomaPatient1__cell_3655 Stroma      -883.1752
#> HumanMelanomaPatient1__cell_3657 Stroma      -894.8463
#> HumanMelanomaPatient1__cell_3658 Stroma      -904.1115
#> HumanMelanomaPatient1__cell_3660 Stroma      -907.8909
#> HumanMelanomaPatient1__cell_3661 Stroma      -874.2712
#> HumanMelanomaPatient1__cell_3663 Stroma      -903.6559

se_results <- SpatialEcoTyper(normdata, scmeta,
                              outprefix = "Melanoma1_subset",
                              radius = 50, ncores = 2)
#> 2024-08-21 14:37:13 Remove 88 genes expressed in fewer than 5 cells
#> 2024-08-21 14:37:13 Construct spatial meta cells for each cell type
#> 		Total spatial microregions: 2315
#> 		Total spatial meta cells: 6343
#> 		Total single cells: 27907
#> 2024-08-21 14:37:14 PCA for each cell type
#> 2024-08-21 14:37:16 Construct cell-type-specific similarity network
#> 		Remove 609 spatial microregions with fewer than 2 cell types 
#> 2024-08-21 14:37:16 Similarity network fusion
#> 2024-08-21 14:37:16 Normalize networks ...
#> 2024-08-21 14:37:17 Calculate the local transition matrix ...
#> 2024-08-21 14:37:17 Perform the diffusion ...
#> 	2024-08-21 14:37:17 Iteration: 1
#> 	2024-08-21 14:37:20 Iteration: 2
#> 	2024-08-21 14:37:30 Iteration: 3
#> 	2024-08-21 14:37:40 Iteration: 4
#> 	2024-08-21 14:37:50 Iteration: 5
#> Warning: The default method for RunUMAP has changed from calling Python UMAP via reticulate to the R-native UWOT using the cosine metric
#> To use Python UMAP via reticulate, set umap.method to 'umap-learn' and metric to 'correlation'
#> This message will be shown once per session
# Extract the Seurat object and updated single-cell metadata
obj <- se_results$obj # A Seurat object
obj
#> An object of class Seurat 
#> 1633 features across 1633 samples within 1 assay 
#> Active assay: RNA (1633 features, 1633 variable features)
#>  2 dimensional reductions calculated: pca, umap
scmeta <- se_results$metadata %>% arrange(SE) # Updated single-cell meta data, with SE annotation added
head(scmeta)
#>                                         X         Y   CellType CellTypeName
#> HumanMelanomaPatient1__cell_4675 4578.019 -6445.993 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_4738 4529.162 -6504.117 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_4774 4533.228 -6529.591 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_4803 4340.813 -6562.381 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_4806 4401.938 -6567.152 Fibroblast  Fibroblasts
#> HumanMelanomaPatient1__cell_4808 4388.422 -6570.215 Fibroblast  Fibroblasts
#>                                        Region Dist2Interface  SE
#> HumanMelanomaPatient1__cell_4675 Outer margin     -197.09807 SE0
#> HumanMelanomaPatient1__cell_4738 Outer margin     -186.13295 SE0
#> HumanMelanomaPatient1__cell_4774 Outer margin     -199.69317 SE0
#> HumanMelanomaPatient1__cell_4803 Outer margin      -95.54956 SE0
#> HumanMelanomaPatient1__cell_4806 Outer margin     -124.01602 SE0
#> HumanMelanomaPatient1__cell_4808 Outer margin     -119.07031 SE0